requirements.txt
certifi==2019.11.28
chardet==3.0.4
feedparser==5.2.1
future==0.18.2
idna==2.8
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.10.3
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.13.0
tweepy==3.8.0
urllib3==1.25.7
